requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

